requirements.txt
#Admin
django-admin-interface==0.16.3

# Django
django==3.2.9

# Django REST Framework
djangorestframework==3.13.1

#Email
sendgrid==6.7.0

django-cors-headers==3.11.0